+++ src/sample/Ingredientes/Bacon.java
@@ -3,7 +3,7 @@ package sample.Ingredientes;
 import sample.Ingrediente;
 
 public class Bacon extends Ingrediente {
-    public Bacon(int precio, String nombre) {
+    public Bacon(double precio, String nombre) {
         super(precio, nombre);
     }
 }
