+++ src/sample/Ingredientes/Lechuga.java
@@ -3,7 +3,7 @@ package sample.Ingredientes;
 import sample.Ingrediente;
 
 public class Lechuga extends Ingrediente {
-    public Lechuga(int precio, String nombre) {
+    public Lechuga(double precio, String nombre) {
         super(precio, nombre);
     }
 }
