+++ src/sample/Ingredientes/Pan.java
@@ -3,7 +3,7 @@ package sample.Ingredientes;
 import sample.Ingrediente;
 
 public abstract class Pan extends Ingrediente{
-    public Pan(int precio, String nombre) {
+    public Pan(double precio, String nombre) {
         super(precio, nombre);
     }
 }
