+++ src/sample/Ingredientes/Panes/Chapata.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Panes;
 import sample.Ingredientes.Pan;
 
 public class Chapata extends Pan {
-    public Chapata(int precio, String nombre) {
+    public Chapata(double precio, String nombre) {
         super(precio, nombre);
     }
 }
