+++ src/sample/Ingredientes/Panes/Hamburguesa.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Panes;
 import sample.Ingredientes.Pan;
 
 public class Hamburguesa extends Pan {
-    public Hamburguesa(int precio, String nombre) {
+    public Hamburguesa(double precio, String nombre) {
         super(precio, nombre);
     }
 }
