+++ src/sample/Ingredientes/Panes/Molde.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Panes;
 import sample.Ingredientes.Pan;
 
 public class Molde extends Pan {
-    public Molde(int precio, String nombre) {
+    public Molde(double precio, String nombre) {
         super(precio, nombre);
     }
 }
