+++ src/sample/Ingredientes/Pepinillos.java
@@ -3,7 +3,7 @@ package sample.Ingredientes;
 import sample.Ingrediente;
 
 public class Pepinillos extends Ingrediente {
-    public Pepinillos(int precio, String nombre) {
+    public Pepinillos(double precio, String nombre) {
         super(precio, nombre);
     }
 }
