+++ src/sample/Ingredientes/Queso.java
@@ -3,7 +3,7 @@ package sample.Ingredientes;
 import sample.Ingrediente;
 
 public class Queso extends Ingrediente {
-    public Queso(int precio, String nombre) {
+    public Queso(double precio, String nombre) {
         super(precio, nombre);
     }
 }
