+++ src/sample/Ingredientes/Quesos/Cabra.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Quesos;
 import sample.Ingredientes.Queso;
 
 public class Cabra extends Queso{
-    public Cabra(int precio, String nombre) {
+    public Cabra(double precio, String nombre) {
         super(precio, nombre);
     }
 }
