+++ src/sample/Ingredientes/Quesos/Cheddar.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Quesos;
 import sample.Ingredientes.Queso;
 
 public class Cheddar extends Queso {
-    public Cheddar(int precio, String nombre) {
+    public Cheddar(double precio, String nombre) {
         super(precio, nombre);
     }
 }
