+++ src/sample/Ingredientes/Salsas/Barbacoa.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Salsas;
 import sample.Ingrediente;
 
 public class Barbacoa extends Ingrediente{
-    public Barbacoa(int precio, String nombre) {
+    public Barbacoa(double precio, String nombre) {
         super(precio, nombre);
     }
 }
