+++ src/sample/Ingredientes/Salsas/Ketchup.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Salsas;
 import sample.Ingrediente;
 
 public class Ketchup extends Ingrediente {
-    public Ketchup(int precio, String nombre) {
+    public Ketchup(double precio, String nombre) {
         super(precio, nombre);
     }
 }
