+++ src/sample/Ingredientes/Salsas/Mayonesa.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Salsas;
 import sample.Ingrediente;
 
 public class Mayonesa extends Ingrediente {
-    public Mayonesa(int precio, String nombre) {
+    public Mayonesa(double precio, String nombre) {
         super(precio, nombre);
     }
 }
