+++ src/sample/Ingredientes/Salsas/Mostaza.java
@@ -3,7 +3,7 @@ package sample.Ingredientes.Salsas;
 import sample.Ingrediente;
 
 public class Mostaza extends Ingrediente {
-    public Mostaza(int precio, String nombre) {
+    public Mostaza(double precio, String nombre) {
         super(precio, nombre);
     }
 }
