+++ src/sample/Ingredientes/Tomate.java
@@ -3,7 +3,7 @@ package sample.Ingredientes;
 import sample.Ingrediente;
 
 public class Tomate extends Ingrediente {
-    public Tomate(int precio, String nombre) {
+    public Tomate(double precio, String nombre) {
         super(precio, nombre);
     }
 }
